+++ lib/moform.dart
@@ -1,3 +1,6 @@
-export 'package:moform/src/widgets/model_int_field.dart';
-export 'package:moform/src/widgets/model_string_field.dart';
 export 'package:moform/src/text_field_builder.dart';
+export 'package:moform/src/widgets/date_field.dart';
+export 'package:moform/src/widgets/date_time_field.dart';
+export 'package:moform/src/widgets/int_field.dart';
+export 'package:moform/src/widgets/string_field.dart';
+export 'package:moform/src/widgets/time_field.dart';